requirements.txt
pathlib
matplotlib
numpy
torch
pyyaml
ipykernel
tqdm